Clean up unused special case python evaluators
