+++ core/src/main/scala/com/raphtory/internals/management/PythonEncoder.scala
@@ -1,28 +1,10 @@
 package com.raphtory.internals.management
 
-import cats.Functor
-import cats.Id
-import cats.syntax.all._
-import com.raphtory.api.analysis.graphstate.Accumulator
-import com.raphtory.api.analysis.graphstate.GraphState
-import com.raphtory.api.analysis.graphview.GraphPerspective
-import com.raphtory.api.analysis.table.Table
-import com.raphtory.api.analysis.visitor.Vertex
-import com.raphtory.api.input.GraphBuilder
-import com.raphtory.internals.management.python.EmbeddedPython
-import com.raphtory.internals.management.python.UnsafeEmbeddedPythonProxy
-import com.typesafe.scalalogging.Logger
-import org.slf4j.LoggerFactory
-
 import java.util
 import scala.collection.mutable
 import scala.reflect.ClassTag
-import scala.util.Random
 import scala.util.Success
 import scala.util.Try
-import scala.jdk.CollectionConverters._
-import scala.reflect.runtime.universe
-import com.raphtory.internals.management.python.UnsafeEmbeddedPythonProxy.global
 
 trait PythonEncoder[A] extends Serializable {
   def encode(a: A): Object
@@ -194,154 +176,4 @@ object PythonEncoder {
 
   implicit def gen[T]: PythonEncoder[T] = macro Magnolia.gen[T]
 
-}
-
-object PythonInterop {
-  val logger: WrappedLogger = new WrappedLogger(Logger(LoggerFactory.getLogger(this.getClass)))
-
-  def assign_id(s: String): Long =
-    GraphBuilder.assignID(s)
-
-  def camel_to_snake(s: String): String =
-    PythonEncoder.camelToSnakeCase(s)
-
-  def testArgs(args: Any): Unit =
-    println(s"testArgs($args)")
-
-  def decode[T](obj: Any): T =
-    (obj match {
-      case obj: java.util.ArrayList[_] => obj.asScala
-      case obj                         => obj
-    }).asInstanceOf[T]
-
-  def get_wrapper_str(obj: Any): String =
-    obj match {
-      case _: Iterable[_]       => "Iterable"
-      case _: Iterator[_]       => "Iterator"
-      case _: GraphPerspective  => "TemporalGraph"
-      case _: Table             => "Table"
-      case _: Accumulator[_, _] => "Accumulator"
-      case _: Vertex            => "Vertex"
-      case _: GraphState        => "GraphState"
-      case _                    => "None"
-    }
-
-  def find_class(name: String): Any = {
-    val runtimeMirror = universe.runtimeMirror(getClass.getClassLoader)
-    val module        = runtimeMirror.staticModule(name)
-    val obj           = runtimeMirror.reflectModule(module)
-    obj.instance
-  }
-
-  def make_varargs(obj: Iterable[Any]): Array[Object] =
-    obj.toArray.map(_.asInstanceOf[Object])
-
-  def methods(obj: Any): util.Map[String, Array[Method]] = {
-    logger.trace(s"Scala 'methods' called with $obj")
-    val prefixedMethodDict         = mutable.Map.empty[String, mutable.ArrayBuffer[java.lang.reflect.Method]]
-    val prefixedMethodDefaultsDict = mutable.Map.empty[String, mutable.Map[Int, java.lang.reflect.Method]]
-    obj.getClass.getMethods.foreach { m =>
-      val parts: Array[String] = """\$default\$""".r.split(m.getName)
-      val prefix               = camel_to_snake(parts(0))
-
-      if (parts.length == 1)
-        prefixedMethodDict.getOrElseUpdate(prefix, mutable.ArrayBuffer.empty[java.lang.reflect.Method]).append(m)
-      else {
-        val defaultIndex = parts(1).toInt - 1
-        prefixedMethodDefaultsDict
-          .getOrElseUpdate(prefix, mutable.Map.empty[Int, java.lang.reflect.Method])
-          .addOne(defaultIndex, m)
-      }
-    }
-    val res                        = prefixedMethodDict
-      .map {
-        case (name, methods) =>
-          val defaults = prefixedMethodDefaultsDict.get(name) match {
-            case Some(v) => v.toMap
-            case None    => Map.empty[Int, java.lang.reflect.Method]
-          }
-
-          name -> methods.map { m =>
-            val hasVarArgs  = m.isVarArgs
-            val paramsNames = m.getParameters.map(p => camel_to_snake(p.getName))
-
-            val n = m.getParameterCount
-            if (
-                    defaults.forall {
-                      case (i, d) => i < m.getParameterCount && m.getParameterTypes()(i) == d.getReturnType
-                    }
-            )
-              Method(m.getName, n, paramsNames, defaults.view.mapValues(_.getName).toMap, hasVarArgs)
-            else
-              Method(m.getName, n, paramsNames, Map.empty[Int, String], hasVarArgs)
-          }.toArray
-      }
-      .toMap
-      .asJava
-    logger.trace(s"Returning found methods for $obj")
-    res
-  }
-
-}
-
-class WrappedLogger(logger: Logger) {
-
-  def level: Int =
-    if (logger.underlying.isTraceEnabled) 5
-    else if (logger.underlying.isDebugEnabled) 4
-    else if (logger.underlying.isInfoEnabled) 3
-    else if (logger.underlying.isWarnEnabled) 2
-    else if (logger.underlying.isErrorEnabled) 1
-    else 0
-
-  def info(msg: String): Unit = logger.info(msg)
-
-  def debug(msg: String): Unit = logger.debug(msg)
-
-  def trace(msg: String): Unit = logger.trace(msg)
-
-  def warn(msg: String): Unit = logger.warn(msg)
-
-  def error(msg: String): Unit = logger.error(msg)
-}
-
-case class Method(name: String, n: Int, parameters: Array[String], defaults: Map[Int, String], varargs: Boolean) {
-  def has_defaults: Boolean = defaults.nonEmpty
-}
-
-/**
-  * Reference of object inside python
-  * @param name
-  * name of variable inside the python context
-  */
-case class PyRef(name: String)
-
-trait PythonFunction {
-  protected val pickleBytes: Array[Byte]
-  protected val eval_name = s"_${Random.alphanumeric.take(32).mkString}"
-
-  @transient lazy val py: EmbeddedPython[Id] = {
-    val _py = UnsafeEmbeddedPythonProxy.global
-    _py.set(s"${eval_name}_bytes", pickleBytes)
-    _py.run(s"import cloudpickle as pickle; $eval_name = pickle.loads(${eval_name}_bytes)")
-    _py.run(s"del ${eval_name}_bytes")
-    _py
-  }
-
-  def invoke(args: Vector[Object]): Id[Object] =
-    py.invoke(PyRef(eval_name), "eval_from_jvm", args)
-}
-
-case class PythonFunction1[I <: AnyRef, R](pickleBytes: Array[Byte]) extends (I => Id[R]) with PythonFunction {
-
-  override def apply(v1: I): Id[R] =
-    invoke(Vector(v1)).map(v => v.asInstanceOf[R])
-}
-
-case class PythonFunction2[I <: AnyRef, J <: AnyRef, R](pickleBytes: Array[Byte])
-        extends ((I, J) => Id[R])
-        with PythonFunction {
-
-  override def apply(v1: I, v2: J): Id[R] =
-    invoke(Vector(v1, v2)).map(_.asInstanceOf[R])
 }
