--- core/src/main/scala/com/raphtory/internals/management/python/PythonIterateEvaluator.scala
@@ -1,21 +1,0 @@
-package com.raphtory.internals.management.python
-
-import cats.syntax.functor._
-import cats.Functor
-import com.raphtory.api.analysis.visitor.Vertex
-import com.raphtory.internals.management.PyRef
-
-import scala.util.Random
-
-class PythonIterateEvaluator[IO[_]: Functor](pyObjBytes: Array[Byte], py: EmbeddedPython[IO])
-        extends (Vertex => IO[Unit]) {
-
-  private val eval_name = s"_${Random.alphanumeric.take(32).mkString}"
-  py.set(s"${eval_name}_bytes", pyObjBytes)
-  py.run(s"import cloudpickle as pickle; $eval_name = pickle.loads(${eval_name}_bytes)")
-  py.run(s"del ${eval_name}_bytes")
-
-  override def apply(v: Vertex): IO[Unit] =
-    py.invoke(PyRef(eval_name), "eval_from_jvm", Vector(v)).map(_ => ())
-
-}
